Support package:// and model:// uri in command line tools
