+++ src/model/src/SolidShapes.cpp
@@ -3,6 +3,7 @@
 
 #include <iDynTree/Model.h>
 #include <iDynTree/SolidShapes.h>
+#include <iDynTree/URIUtils.h>
 
 #include <cstdlib>
 #include <fstream>
@@ -279,115 +280,7 @@ const std::vector<std::string>& ExternalMesh::getPackageDirs() const
 
 std::string ExternalMesh::getFileLocationOnLocalFileSystem() const
 {
-    bool isWindows = false;
-#ifdef _WIN32
-    isWindows = true;
-#endif
-
-    std::unordered_set<std::string> pathList;
-
-    // if the user provides the packageDirs we do not check inside the dirs listed in the
-    // environment variables
-    if (!this->packageDirs.empty())
-    {
-        pathList
-            = std::unordered_set<std::string>(this->packageDirs.begin(), this->packageDirs.end());
-    } else
-    {
-        // List of variables that contain <prefix>/share paths
-        std::vector<std::string> envListShare
-            = {"GZ_SIM_RESOURCE_PATH", "GAZEBO_MODEL_PATH", "ROS_PACKAGE_PATH"};
-        // List of variables that contains <prefix> paths (to which /share needs to be added)
-        std::vector<std::string> envListPrefix = {"AMENT_PREFIX_PATH"};
-
-        for (size_t i = 0; i < envListShare.size(); ++i)
-        {
-            const char* env_var_value = std::getenv(envListShare[i].c_str());
-
-            if (env_var_value)
-            {
-                std::stringstream env_var_string(env_var_value);
-
-                std::string individualPath;
-
-                while (std::getline(env_var_string, individualPath, isWindows ? ';' : ':'))
-                {
-                    pathList.insert(individualPath);
-                }
-            }
-        }
-
-        for (size_t i = 0; i < envListPrefix.size(); ++i)
-        {
-            const char* env_var_value = std::getenv(envListPrefix[i].c_str());
-
-            if (env_var_value)
-            {
-                std::stringstream env_var_string(env_var_value);
-
-                std::string individualPath;
-
-                while (std::getline(env_var_string, individualPath, isWindows ? ';' : ':'))
-                {
-                    pathList.insert(individualPath + "/share");
-                }
-            }
-        }
-    }
-    auto cleanPathSeparator = [isWindows](const std::string& filename) -> std::string {
-        std::string output = filename;
-        char pathSeparator = isWindows ? '\\' : '/';
-        char wrongPathSeparator = isWindows ? '/' : '\\';
-
-        for (size_t i = 0; i < output.size(); ++i)
-        {
-            if (output[i] == wrongPathSeparator)
-            {
-                output[i] = pathSeparator;
-            }
-        }
-
-        return output;
-    };
-
-    auto isFileExisting = [](const std::string& filename) -> bool {
-        if (FILE* file = fopen(filename.c_str(), "r"))
-        {
-            fclose(file);
-            return true;
-        } else
-        {
-            return false;
-        }
-    };
-
-    auto getFilePath
-        = [isFileExisting, cleanPathSeparator](const std::string& filename,
-                                               const std::string& prefixToRemove,
-                                               const std::unordered_set<std::string>& paths) {
-              if (isFileExisting(filename))
-              {
-                  return filename;
-              }
-
-              if (filename.substr(0, prefixToRemove.size()) == prefixToRemove)
-              {
-                  std::string filename_noprefix = filename;
-                  filename_noprefix.erase(0, prefixToRemove.size());
-                  for (const std::string& path : paths)
-                  {
-                      const std::string testPath = cleanPathSeparator(path + filename_noprefix);
-                      if (isFileExisting(testPath))
-                      {
-                          return testPath;
-                      }
-                  }
-              }
-
-              return filename; // By default return the input;
-          };
-
-    return getFilePath(filename, "package:/", pathList);
+    return resolveURI(filename, packageDirs);
 }
 
 void ExternalMesh::setFilename(const std::string& filename)
